[fix] Long 타입 오류 해결

+++ src/main/kotlin/com/photi/server/domain/report/custom/ReportCategoryCustomRepository.kt
@@ -7,5 +7,5 @@ interface ReportCategoryCustomRepository {
 
     fun findAllDescription(type: ReportCategoryType?): List<String>
 
-    fun find(id: Int, type: ReportCategoryType?): ReportCategory?
+    fun find(id: Long, type: ReportCategoryType?): ReportCategory?
 }
